SSR safe useMutationObserver
Our components could be rendered in the server, where `MutationObserver` is not available. If that's the case, don't call `new` on it.

Test Plan:
Load testbed, check that the mutation observer is called with the latest callback I give to it. (Added some code to conditionally pass a different callback based on some state, and a button to change that state).


Reviewers: flooey, lazybean, elenaperezrioja, mhtien, Netceer

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/9121

monorepo-commit: b54de0c0ddb654fe9839ed41ac6a528af099a126