Message: memoize element passed to layout (#9134)
Summary:
Despite those elements passed to the MessageLayout not re-rendering, MessageLayout think they are changing and will rererender itself. So we memoize element passed to layout


It looks nicer in profiling.
I do not expect to have much effect on perf, but it cleans up the profile and should make bad rerender more obvious.



![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/IoyKnJXsDnJREClofLDY/251c5e8d-e488-4ad6-80d2-455d46d3e4a0.png)

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/IoyKnJXsDnJREClofLDY/6931c3b2-c783-4583-b236-c8973cb536db.png)


Test Plan:

Reviewers:

monorepo-commit: 5ee393fecc76ea3eb8e535bebe11cf4a1ee56848