Update ThreadAddButtonToMentionMenu with VirtualizedList
When speaking with Ludo we noticed the example for
adding an email invite link did not show up in the docs.

This was due to the recent changes to moving the
mentions list into a VirtualizedList so after
discussing with Elena, I've updated the docs by
replacing the VirtualizedList with a Menu component.

I've made the mention list an item of the new Menu
as I wanted the invite link to always appear
at the bottom. This meant also changing the styles
of cord-mention-menu (border, margin etc) so it didn't
look like a nested menu.

Also had to export the VirtualizedList to use it.


Test Plan:
Refer to screen shot
![image](https://github.com/getcord/monorepo/assets/59407030/0c6857ae-9e5b-4d3c-a99d-bfc884b9ffaf)

Reviewers: AlbertQM, elenaperezrioja, lazybean, Netceer

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/9132

monorepo-commit: 4a92dbd572b1345144363775261396eb75559727