magic `useCordIDs` hook (#9126)
Summary:
When replacing components you may want to know what thread/message you
ar in, what the user (author, or avatar) is.
But you cannot, because we are not passing this as prop everywhere.

The easy but cumbersome solution was to replace the parent
thread/messag/avatar/... with a wrapped in Context Provider version of
itself.

Now just use `useCordIDs`, you ll get what you need.

Test Plan:

Reviewers:

monorepo-commit: 2c9aaf3e31fc9dbdf08796c242c0a88513f220a0