Add InlineThreadHeader component to Threads
Can be controlled on Threads and InlineThread
levels.

Default is to hide headers but if they are enabled
then it relies on the thread name and url to
populate the header title and href.

This does introduce border styling to all inline threads.


Reviewers: lazybean, elenaperezrioja, mhtien, AlbertQM

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/9120

monorepo-commit: 9c1d4db17f8a930a38fbde02fbf5b147e8a5104c