Remove setting message type from JS API (#9164)
In order to eliminate a way for users to possibly deceive others about
what's going on, remove the ability to create action messages from the
JS API.  They can still be created via the REST API if needed.


Test Plan:

Attempted to create an action message using the JS API, it came
through as a user message.

Sent a message using a beta component, it came through correctly.

Sent a message using an onion component, it came through correctly.

monorepo-commit: 20dfd7ea5c5636797fc53fdec7bf0b85a12420ec