Change markdown format to be less antagonistic to Slate
Slate seems to like to distinguish between a "block" element, with a
type and children, and an "inline" text element (with no type). Having a
combo of the two here is fine for our rendering, but is going to be
problematic to edit with the Slate composer if/when we eventually get
there.

Given that this is undocumented and has very incomplete support, now
seems the time to change it.


Reviewers: flooey, elenaperezrioja, dmmiller

Reviewed By: flooey, dmmiller

Pull Request: https://github.com/getcord/monorepo/pull/9156

monorepo-commit: fd0d5021b62153ae956550695393b00385e194d9