Create and add EmptyThreadsPlaceholder to Threads
Noticed we didn't have one so adding it
using the EmptyPlaceholder.

Currently I've just added the placeholder users
if a groupID exists in the composerOptions.

Not sure what to do if users do not state a groupID,
will have a think about it.


Test Plan: Tested locally and can see the placeholder!

Reviewers: elenaperezrioja, lazybean, AlbertQM, flooey, Netceer

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/9101

monorepo-commit: 5dc500a1956d5313ec9ecc77ee01ba6584c417c3