Fix useUpdatingRef (#9172)
The previous version of useUpdatingRef updated the ref in a useEffect,
which meant that in the current render cycle (when useUpdatingRef
returns), the new value wasn't available in the ref.  A bunch of our
code relied on the new value being available, so move back to just
setting the ref immediately.

Test Plan:

The `fetchMore` callback from observeThreads was reissuing an old
request for more when it was called in the testbed, after this diff it
doesn't.

monorepo-commit: 4ca57718b21690b62605c3ef5b791dcb9d11093e