Add replace prop to tabbedThreadsOptions
Realised as I was building the unresolved/resolved
threads with tabs I couldn't replace a specific
component within only one of the Threads component
within the tabs.

Then Ludo had the great idea of passing in a replace
prop so this is the PR for that!


Test Plan: Will be used in #9149

Reviewers: elenaperezrioja, dmmiller, lazybean, Netceer

Reviewed By: dmmiller

Pull Request: https://github.com/getcord/monorepo/pull/9186

monorepo-commit: e425e02ce1206b531d31db5eda3e2b923d35c7ea