ScrollContainer:Observe resize of children (#9163)
Summary:
One issue with current ScrollContainer is that if a child is changing
size, it wont scroll or check overflow.

A typical example is message with images.
When image is loading the message is smaller.

The solution was proposed by @AlbertQM , observing resize of the children.

Test Plan:
2 browsers with test bed and different users.
One one scroll to bottom of thread.
On the other send message with copypaste very high screenshot (*2).
On the first one, new message is fully visible.
 
Reviewers:

monorepo-commit: 691a40662342789a6215fd1871fb5c1a1b78ad90