Improve message sending pathway (#9219)
Now that we have optimistic updates of the API for sending a message,
message sending in the new components feels janky because the message
stays in the composer until it's actually sent, but it *also* shows up
in the message list immediately.  I think we can do better.

Instead of doing that, clear the composer immediately when the message
is sent, but stash the previous value of the composer.  If the sending
succeeds, great, nothing to do.  If it fails, give the user a link to
restore their message to the composer.  (The text there could probably
be improved, but I couldn't think of anything better.)  I decided to
include a link rather than restoring it immediately because the user
might have already started to write their next message, and I wouldn't
want to overwrite that.

Test Plan:

Added a delay of 3s to sending messages on my local server, then tried
various things:

Sent a message normally.  It disappeared from the composer immediately
and appeared in the thread list.  I could send messages as fast as I
could type.

Added `initialValue={{ id: '123' }}` to the composer, which meant that
any message after the first would fail due to reusing an ID.  The
second and subsequent failures caused the message to disappear from
the thread list and show the error.  Clicking the link on the error
restored the message to the composer.

Hit enter a bunch of times really quickly while the message was
sending.  I still ended up with a single copy of the message.  (This
used to be a problem.)

monorepo-commit: e2a76630bfee48ee4e3b16742b883bf12a3c9399