Update dayjs file and stabilize order (#9237)
The dayjs.ts file needed to be updated because we updated the dayjs
library, but the locale.json file from dayjs had the locales reordered
for some reason, which added a bunch of churn to the dayjs.ts file.
Sort the list of keys first before generating the output, and then
regenerate the file, so it will be more stable across releases.

Test Plan:

Mark I Eyeball

monorepo-commit: 78b36669a483e5e5bf70bbfa614197c9f21785ad