Add screenshot taking to API (#9235)
Add the ability to capture a screenshot at message send time to the JS
API.  This doesn't exist in the REST API because the screenshot
functionality is done in the browser automatically and there's no
equivalent on the REST API.

The type of `addScreenshot` is a boolean right now, but if we want to
add options, we can allow it to be an object as well and replace
`true` with the default option values to maintain backwards
compatibility.  I considered having it be an object from the start,
but an API where you have to supply `addScreenshot: {}` is very
awkward, so accepting `true` seemed like the right choice.

Test Plan:

Send a message from the JS API specifying `{addScreenshot: true}` .
The email notification contained a screenshot.

monorepo-commit: 2d49f4eadf1e22e72b7e0fcc76979c13611d0de3